Refactor. Add global styles for heading and quote
