+++ lib/demo-block-templates/front-page.html
@@ -28,6 +28,16 @@
 	scelerisque. Etiam euismod blandit leo a maximus.</p>
 <!-- /wp:paragraph -->
 
+<!-- wp:quote -->
+<blockquote class="wp-block-quote">
+	<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit</p>
+</blockquote>
+<!-- /wp:quote -->
+
+<!-- wp:button -->
+<div class="wp-block-button"><a class="wp-block-button__link">Hello</a></div>
+<!-- /wp:button -->
+
 <!-- wp:columns {"align":"wide"} -->
 <div class="wp-block-columns alignwide">
 	<!-- wp:column -->
