+++ packages/block-library/src/quote/style.scss
@@ -17,3 +17,9 @@
 		}
 	}
 }
+
+.wp-gs .wp-block-quote {
+	p {
+		font-size: var(--wp-quote--fontSize);
+	}
+}
