+++ packages/edit-site/src/components/sidebar/global-styles-panel/panels/index.js
@@ -1,0 +1,4 @@
+export { default as ColorPanel } from './color-panel';
+export { default as HeadingPanel } from './heading-panel';
+export { default as TypographyPanel } from './typography-panel';
+export { default as QuotePanel } from './quote-panel';
